+++ yuio/string.py
@@ -3472,7 +3472,7 @@ class Link(_StrBase):
         res = ColorizedString()
         res.start_link(self._url)
         res.append_colorized_str(ctx.str(self._msg))
-        if not ctx.term.supports_colors:
+        if not ctx.term.supports_colors or yuio.term.detect_ci():
             res.start_no_wrap()
             res.append_str(" [")
             res.append_str(self._url)